popdoc: split popdoc.js into src/ modules
